some tests added, some tests still failing
