+++ tests/pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = -x -p no:warnings --cov-report=term-missing --cov-report=term --cov-report=xml:./coverage.xml --no-cov-on-fail --cov=../fastapi_auth_middleware
+;addopts = -x -p no:warnings --cov-report=term-missing --cov-report=term --cov-report=xml:./coverage.xml --no-cov-on-fail --cov=../fastapi_auth_middleware
